新增 - 【招聘】创建附件 - 【招聘】获取招聘官网下的职位列表 - 【招聘】获取项目列表 - 【招聘】获取职能分类列表 - 【招聘】搜索招聘官网下的职位列表 - 【招聘】创建人才 - 【招聘】获取申请表附加信息 - 【招聘】删除备注
变更
- 时间戳类型 int=>long

+++ src/Hire/GetHireV1InterviewsResponseDto.cs
@@ -4,7 +4,7 @@
 // Created          : 2024-06-24
 //
 // Last Modified By : yxr
-// Last Modified On : 2024-06-24
+// Last Modified On : 2024-07-09
 // ************************************************************************
 // <copyright file="GetHireV1InterviewsResponseDto.cs" company="Vicente Yu">
 //     MIT
@@ -47,7 +47,7 @@ public record GetHireV1InterviewsResponseDto
         /// <para>示例值：1618899376474</para>
         /// </summary>
         [JsonPropertyName("begin_time")]
-        public int? BeginTime { get; set; }
+        public long? BeginTime { get; set; }
 
         /// <summary>
         /// <para>面试结束时间，毫秒时间戳</para>
@@ -55,7 +55,7 @@ public record GetHireV1InterviewsResponseDto
         /// <para>示例值：1618999376474</para>
         /// </summary>
         [JsonPropertyName("end_time")]
-        public int? EndTime { get; set; }
+        public long? EndTime { get; set; }
 
         /// <summary>
         /// <para>面试轮次</para>
